_data_scripts/requirements.txt
python-slugify
pyyaml
requests
